+++ lib/ui/component/error_message.dart
